+++ Frameworks/OpenMPT/OpenMPT/soundlib/ContainerXPK.cpp
@@ -98,6 +98,8 @@ static uint8 XPK_ReadTable(int32 index)
 	};
 	if(index < 0) throw XPK_error();
 	if(static_cast<std::size_t>(index) >= std::size(xpk_table)) throw XPK_error();
+	// cppcheck false-positive
+	// cppcheck-suppress arrayIndexOutOfBoundsCond
 	return xpk_table[index];
 }