+++ Frameworks/OpenMPT/OpenMPT/soundlib/Load_dsym.cpp
@@ -258,14 +258,16 @@ bool CSoundFile::ReadDSym(FileReader &file, ModLoadingFlags loadFlags)
 
 	const auto allowedCommands = file.ReadArray<uint8, 8>();
 
-	std::vector<std::byte> sequenceData;
+	std::vector<uint16le> sequence;
 	if(fileHeader.numOrders)
 	{
+		std::vector<std::byte> sequenceData;
 		const uint32 sequenceSize = fileHeader.numOrders * fileHeader.numChannels * 2u;
 		if(!ReadDSymChunk(file, sequenceData, sequenceSize))
 			return false;
+		FileReader sequenceChunk = FileReader(mpt::as_span(sequenceData));
+		sequenceChunk.ReadVector(sequence, sequenceData.size() / 2u);
 	}
-	const auto sequence = mpt::as_span(reinterpret_cast<uint16le *>(sequenceData.data()), sequenceData.size() / 2u);
 
 	std::vector<std::byte> trackData;
 	trackData.reserve(fileHeader.numTracks * 256u);