+++ Frameworks/OpenMPT/OpenMPT/soundlib/S3MTools.h
@@ -50,6 +50,7 @@ struct S3MFileHeader
 		trkIT2_07         = 0x3207,
 		trkIT2_14         = 0x3214,
 		trkBeRoTrackerOld = 0x4100,  // Used from 2004 to 2012
+		trkGraoumfTracker = 0x5447,
 		trkCamoto         = 0xCA00,
 	};