+++ Frameworks/OpenMPT/OpenMPT/soundlib/Sndfile.cpp
@@ -1074,6 +1074,7 @@ PlayBehaviourSet CSoundFile::GetSupportedPlaybackBehaviour(MODTYPE type)
 		playBehaviour.set(kITDoNotOverrideChannelPan);
 		playBehaviour.set(kITDCTBehaviour);
 		playBehaviour.set(kITPitchPanSeparation);
+		playBehaviour.set(kITResetFilterOnPortaSmpChange);
 		if(type == MOD_TYPE_MPT)
 		{
 			playBehaviour.set(kOPLFlexibleNoteOff);