+++ Frameworks/OpenMPT/OpenMPT/src/mpt/format/tests/tests_format_simple.hpp
@@ -14,6 +14,7 @@
 #include "mpt/test/test_macros.hpp"
 
 #include <limits>
+#include <string>