+++ Frameworks/OpenMPT/OpenMPT/src/mpt/io_write/buffer.hpp
@@ -8,6 +8,7 @@
 #include "mpt/base/memory.hpp"
 #include "mpt/base/namespace.hpp"
 #include "mpt/base/integer.hpp"
+#include "mpt/base/span.hpp"
 #include "mpt/io/base.hpp"
 
 #include <cstddef>